refactor(frontend): split SettingsContext into 5 domain contexts and extract SettingsPage sections
- Split monolithic SettingsContext into 5 narrow domain contexts
  (GeneralSettings, AppearanceSettings, TaskConfiguration,
  ProjectDefaults, LlmProviders) with a composite provider.
- Extract all 10 inline sections from SettingsPage (1754 → 207 lines)
  into independent components under features/settings/components/.
- Add settingsHelpers.ts for shared utilities.
- Migrate 3 consumers to domain hooks:
  App.tsx → useGeneralSettings, LlmProvidersTab → useLlmProviders,
  TaskConfigurationSection → useLlmProviders.
- Keep legacy useSettings hook for remaining orchestration consumers.


- Create GeneralSettings, AppearanceSettings, TaskConfiguration,
  ProjectDefaults, and LlmProviders contexts with narrow hooks.
- Composite SettingsProvider nests all 5 at once; legacy useSettings
  shim kept via LegacySettingsContext for backwards compatibility.
- Move settings implementation files into features/settings/.
- Keep src/settings/index.ts as temporary re-export barrel.

Verified: npm run lint, tsc -b, npm run test:run, npm run build.
