+++ docs/LLM-Working/worklog/2026-06-15.md
@@ -8,3 +8,11 @@
 - Added `Card`, `CardHeader`, `CardBody`, `cn`, and cva-based Button variants while preserving the existing visual tokens and public Button API.
 - Rewrote consumers/tests to import from `@design-system/primitives`, `@design-system/layout`, and `@design-system/tokens/theme`; kept legacy `src/components/ui` and `src/components/layout` barrels as temporary re-exports.
 - Validation: `npm run lint`, `node_modules/.bin/tsc -b`, `npm run test:run` (190 tests), and `npm run build` all passed from `frontend/`.
+
+## 08:30 [phase1] SettingsContext domain split
+
+- Split monolithic SettingsContext into 5 domain contexts: GeneralSettings, AppearanceSettings, TaskConfiguration, ProjectDefaults, LlmProviders.
+- Each domain context exposes a narrow hook alongside its provider; composite SettingsProvider nests all 5 and also provides a legacy useSettings shim for backwards compatibility.
+- Moved settings implementation files into features/settings/{contexts,utils,hooks,components} and kept src/settings/index.ts as temporary re-export.
+- Fixed type imports that were incorrectly rewritten (chat types, SkillMode) during the earlier sed batch.
+- Validation: npm run lint, tsc -b, npm run test:run (190 tests), npm run build all passed from frontend/.
