+++ frontend/src/design-system/primitives/SkillToggleGroup.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useState } from 'react';
 import type { SkillItem } from '@/shared/types';
-import type { SkillMode } from '../../settings/types';
+import type { SkillMode } from '@features/settings/types';
 import { useT } from '@/shared/i18n';
 
 interface Props {
