+++ frontend/src/features/settings/index.ts
@@ -1,2 +1,46 @@
-// Temporary re-export until settings migration is complete
-export * from '@/settings';
+export {
+  clampEditorFontSize,
+  clampTerminalFontSize,
+  createDefaultTaskConfigurationSettings,
+  createDefaultWebUiSettings,
+  createEmptyEnvironmentTaskDefaults,
+  defaultEditorFontFamily,
+  defaultEditorFontSize,
+  defaultResearchAgentProfileId,
+  defaultTerminalFontSize,
+  maxEditorFontSize,
+  maxTerminalFontSize,
+  minEditorFontSize,
+  minTerminalFontSize,
+  rawPromptTaskConfigurationId,
+  settingsStorageKey,
+  structuredResearchTaskConfigurationId,
+} from './utils/defaults';
+export {
+  SettingsProvider,
+  useEditorSettings,
+  useProjectEnvironmentDefaults,
+  useSettings,
+  useTerminalFontSize,
+} from './contexts/SettingsProvider';
+export { useGeneralSettings } from './contexts/GeneralSettingsContext';
+export { useAppearanceSettings } from './contexts/AppearanceSettingsContext';
+export { useTaskConfiguration } from './contexts/TaskConfigurationContext';
+export { useProjectDefaults } from './contexts/ProjectDefaultsContext';
+export { useLlmProviders } from './contexts/LlmProvidersContext';
+export { readStoredSettings, resolveProjectEnvironmentDefaults } from './utils/storage';
+export type {
+  DefaultRoute,
+  DefaultProjectSelectionState,
+  DefaultProjectSettings,
+  EnvironmentTaskDefaults,
+  ExecutionEngineId,
+  LlmProvider,
+  LlmProviderFormat,
+  ResearchAgentProfileSettings,
+  SettingsRecoveryReason,
+  TaskConfigurationMode,
+  TaskConfigurationPreset,
+  TaskConfigurationSettings,
+  WebUiSettingsDocument,
+} from './types';
