+++ frontend/src/settings/index.ts
@@ -1,35 +1,2 @@
-export {
-  clampEditorFontSize,
-  clampTerminalFontSize,
-  createDefaultTaskConfigurationSettings,
-  createDefaultWebUiSettings,
-  createEmptyEnvironmentTaskDefaults,
-  defaultEditorFontFamily,
-  defaultEditorFontSize,
-  defaultResearchAgentProfileId,
-  defaultTerminalFontSize,
-  maxEditorFontSize,
-  maxTerminalFontSize,
-  minEditorFontSize,
-  minTerminalFontSize,
-  rawPromptTaskConfigurationId,
-  settingsStorageKey,
-  structuredResearchTaskConfigurationId,
-} from './defaults';
-export { SettingsProvider, useEditorSettings, useProjectEnvironmentDefaults, useSettings, useTerminalFontSize } from './context';
-export { readStoredSettings, resolveProjectEnvironmentDefaults } from './storage';
-export type {
-  DefaultRoute,
-  DefaultProjectSelectionState,
-  DefaultProjectSettings,
-  EnvironmentTaskDefaults,
-  ExecutionEngineId,
-  LlmProvider,
-  LlmProviderFormat,
-  ResearchAgentProfileSettings,
-  SettingsRecoveryReason,
-  TaskConfigurationMode,
-  TaskConfigurationPreset,
-  TaskConfigurationSettings,
-  WebUiSettingsDocument,
-} from './types';
+// Temporary re-export until all consumers migrate to @features/settings
+export * from '@features/settings';
